// ClientApp/src/app/core/models/ISimulator.ts
import { IPaymentPlan } from "./IPaymentPlan";
export interface ISimulator {
  projectId: number;
  projectName: string;
  projectToken: string;
  financialCalculationMethod: string;
  segmentId: number;
  segmentName: string;
  monthsUntilSimulationEnd: number;
  monthsUntilProposedSimulationEnd: number;
  buildingId: number;
  buildingName: string;
  buildingArea: number;
  buildingValue: number;
  parkingId: number;
  parkingName: string;
  parkingValue: number;
  storageId: number;
  storageName: string;
  storageValue: number;
  otherId: number;
  otherName: string;
  otherValue: number;
  companyLogo: string;
  companyName: string;
  projectLogo: string;
  projectPrivacyLink: string;
  financingMonths: number;
  favorInterestRate: number;
  defaultInterestRate: number;
  separationFeeValue: number;
  separationFeeProposedValue: number;
  initialFeePercentage: number;
  financingFeePercentage: number;
  initialValue(): number;
  initialProposedValue: number;
  financingValue(): number;
  financingProposedValue: number;
  financingProposedRate: number;
  financingProposedYears: number;
  simulationInterestValue: number;
  originalTotalValue: number;
  totalValue: number;
  projectPaymentPlan: IPaymentPlan[];
  clientPaymentPlan: IPaymentPlan[];
  paymentPlanId: string;
}